Strip out some left-over multifile logic
