+++ inst/editor/src/r-parsing/index.ts
@@ -3,7 +3,6 @@ export type Primatives = string | number | boolean;
 export { generateRAppScriptTemplate } from "./generate_app_script_template";
 export { generate_r_output_binding } from "./generate_output_binding";
 
-export { findUiDefInRApp as get_ui_node_from_r_multifile_app } from "./parse_multifile_r_apps";
 // export { get_server_positions } from "./get_server_positions";
 // export { find_ui_and_server_in_singlefile_app } from "./find_ui_and_server_in_singlefile_app";
 // export { find_ui_and_server_in_multifile_r_app } from "./parse_multifile_r_apps";
